added api endpoints, need to add to frontend still
