+++ src/firebaseconfig.json
@@ -1,0 +1,1 @@
+{"apiKey":"AIzaSyBxg_tzkZg9y_GJdq7DuOU5-NRI12UbD5M","authDomain":"cards-against-humanity-76d45.firebaseapp.com","projectId":"cards-against-humanity-76d45","storageBucket":"cards-against-humanity-76d45.appspot.com","messagingSenderId":"468637051485","appId":"1:468637051485:web:ae1e184b247ee02ff5bbb4","measurementId":"G-6G066019LS"}
